// backend/src/types/enums/genres.ts
export enum Genre {
  Comedy = 'Comedy',
  Fantasy = 'Fantasy',
  Crime = 'Crime',
  Drama = 'Drama',
  Music = 'Music',
  Adventure = 'Adventure',
  History = 'History',
  Thriller = 'Thriller',
  Animation = 'Animation',
  Family = 'Family',
  Mystery = 'Mystery',
  Biography = 'Biography',
  Action = 'Action',
  FilmNoir = 'FilmNoir',
  Romance = 'Romance',
  SciFi = 'SciFi',
  War = 'War',
  Western = 'Western',
  Horror = 'Horror',
  Musical = 'Musical',
  Sport = 'Sport',
}